+++ CHANGES.md
@@ -5,6 +5,7 @@ v0.0.3, 2014.01.29
 ------------------
 
 * Improved `__repr__` for: IncomingMessage, OutgoingMessage, MessageStatus
+* `LoggerProvider`: now allows to omit the `logger` config parameter to use a default logger
 
 v0.0.2, 2014.01.29
 ------------------